--- a/staging/src/k8s.io/apiserver/go.mod
+++ b/staging/src/k8s.io/apiserver/go.mod
@@ -17,7 +17,7 @@ require (
 	github.com/fsnotify/fsnotify v1.7.0
 	github.com/go-logr/logr v1.4.2
 	github.com/gogo/protobuf v1.3.2
-	github.com/google/btree v1.0.1
+	github.com/google/btree v1.1.3
 	github.com/google/cel-go v0.22.0
 	github.com/google/gnostic-models v0.6.9
 	github.com/google/go-cmp v0.6.0