--- a/staging/src/k8s.io/apiserver/pkg/storage/cacher/store.go
+++ b/staging/src/k8s.io/apiserver/pkg/storage/cacher/store.go
@@ -19,8 +19,6 @@ package cacher
 import (
 	"fmt"
 
-	"github.com/google/btree"
-
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -96,12 +94,6 @@ type storeElement struct {
 	Labels labels.Set
 	Fields fields.Set
 }
-
-func (t *storeElement) Less(than btree.Item) bool {
-	return t.Key < than.(*storeElement).Key
-}
-
-var _ btree.Item = (*storeElement)(nil)
 
 func storeElementKey(obj interface{}) (string, error) {
 	elem, ok := obj.(*storeElement)