@wojtek-t 2024-11-20
This is great.

We're in code freeze now, so this won't merge over the next few weeks, but let's queue it for merging as soon as we're out of code freeze in December.

/lgtm
/approve
/milestone v1.33