@dims 2024-11-20
No issue with the dependency update itself!

/approve