@BenTheElder 2024-11-20
So yeah dep-approvers is still broken :/ @thockin @liggitt @dims 

We will have to get regex approve working when we ... have more than n=1 prow maintainers again