@serathius 2024-11-22
@nayihz Feel free to run the benchmarks listed in https://github.com/kubernetes/kubernetes/pull/126754