@serathius 2025-01-03
Run benchstat on the provided benchmark results:
```
goos: darwin
goarch: arm64
pkg: k8s.io/apiserver/pkg/storage/cacher
cpu: Apple M2 Pro
                                                  │ interface.txt │             generic.txt              │            generics.txt             │
                                                  │    sec/op     │    sec/op      vs base               │   sec/op     vs base                │
StoreListCreate/RV=NotOlderThan-10                   216.8µ ±  1%    223.0µ ±  5%       ~ (p=0.247 n=10)   250.7µ ± 2%  +15.63% (p=0.000 n=10)
StoreListCreate/RV=ExactMatch-10                     312.9µ ±  1%    312.7µ ±  4%       ~ (p=0.684 n=10)   377.3µ ± 1%  +20.59% (p=0.000 n=10)
StoreList/RV=Empty-10                                631.9µ ±  3%    636.5µ ±  5%       ~ (p=0.739 n=10)   707.7µ ± 4%  +11.99% (p=0.001 n=10)
StoreList/RV=NotOlderThan-10                         605.8µ ±  7%    597.9µ ±  7%       ~ (p=0.436 n=10)   676.5µ ± 6%  +11.66% (p=0.000 n=10)
StoreList/RV=MatchExact-10                           18.34m ±  2%    18.36m ±  2%       ~ (p=0.796 n=10)   22.82m ± 3%  +24.48% (p=0.000 n=10)
StoreList/Paginate/RV=Empty-10                       20.31m ±  1%    20.18m ±  1%       ~ (p=0.218 n=10)   24.15m ± 1%  +18.90% (p=0.000 n=10)
StoreList/Paginate/RV=NotOlderThan-10                20.28m ±  1%    20.20m ±  1%       ~ (p=0.796 n=10)   24.34m ± 1%  +20.05% (p=0.000 n=10)
StoreList/Paginate/RV=MatchExact-10                  20.29m ±  5%    20.39m ±  1%       ~ (p=0.971 n=10)   24.68m ± 1%  +21.63% (p=0.000 n=10)
StoreList/NodeIndexed/RV=Empty-10                    2.231m ±  3%    2.229m ±  2%       ~ (p=1.000 n=10)   2.561m ± 2%  +14.76% (p=0.000 n=10)
StoreList/NodeIndexed/RV=NotOlderThan-10             1.453m ±  1%    1.464m ±  2%       ~ (p=0.393 n=10)   1.672m ± 2%  +15.10% (p=0.000 n=10)
StoreList/NodeIndexed/RV=MatchExact-10                7.905 ±  1%     7.763 ±  1%  -1.80% (p=0.003 n=10)   10.965 ± 0%  +38.71% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=Empty-10           96.55m ± 10%    99.85m ± 24%       ~ (p=0.315 n=10)   85.12m ± 1%  -11.84% (p=0.002 n=10)
StoreList/NodeIndexed/Paginate/RV=NotOlderThan-10    96.99m ±  5%   101.92m ±  2%  +5.09% (p=0.002 n=10)   84.72m ± 1%  -12.65% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=MatchExact-10       8.133 ±  1%     8.050 ±  2%       ~ (p=0.075 n=10)   11.420 ± 1%  +40.41% (p=0.000 n=10)
StoreList/Namespace/RV=Empty-10                      2.204m ± 12%    2.141m ± 14%       ~ (p=0.529 n=10)   2.441m ± 3%        ~ (p=0.075 n=10)
StoreList/Namespace/RV=NotOlderThan-10               1.425m ±  6%    1.447m ±  5%       ~ (p=0.912 n=10)   1.552m ± 3%   +8.89% (p=0.000 n=10)
StoreList/Namespace/RV=MatchExact-10                 25.82m ±  3%    26.22m ±  8%       ~ (p=0.353 n=10)   22.75m ± 1%  -11.89% (p=0.000 n=10)
geomean                                              11.14m          11.20m        +0.47%                  12.59m       +12.96%

                                                  │ interface.txt │             generic.txt              │             generics.txt             │
                                                  │  objects/op   │ objects/op   vs base                 │ objects/op   vs base                 │
StoreListCreate/RV=NotOlderThan-10                     1.000 ± 0%    1.000 ± 0%       ~ (p=1.000 n=10) ¹    1.000 ± 0%       ~ (p=1.000 n=10) ¹
StoreListCreate/RV=ExactMatch-10                       1.000 ± 0%    1.000 ± 0%       ~ (p=1.000 n=10) ¹    1.000 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/RV=Empty-10                                 10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/RV=NotOlderThan-10                          10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/RV=MatchExact-10                            9.928k ± 0%   9.929k ± 0%       ~ (p=0.154 n=10)     9.924k ± 0%  -0.05% (p=0.000 n=10)
StoreList/Paginate/RV=Empty-10                        10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Paginate/RV=NotOlderThan-10                 10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Paginate/RV=MatchExact-10                   9.928k ± 0%   9.928k ± 0%       ~ (p=1.000 n=10)     9.922k ± 0%  -0.06% (p=0.000 n=10)
StoreList/NodeIndexed/RV=Empty-10                     10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/RV=NotOlderThan-10              10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/RV=MatchExact-10                9.901k ± 0%   9.901k ± 0%       ~ (p=1.000 n=10) ¹   9.901k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/Paginate/RV=Empty-10            10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/Paginate/RV=NotOlderThan-10     10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/Paginate/RV=MatchExact-10       9.901k ± 0%   9.901k ± 0%       ~ (p=1.000 n=10) ¹   9.901k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Namespace/RV=Empty-10                       10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Namespace/RV=NotOlderThan-10                10.00k ± 0%   10.00k ± 0%       ~ (p=1.000 n=10) ¹   10.00k ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Namespace/RV=MatchExact-10                  9.927k ± 0%   9.927k ± 0%       ~ (p=0.898 n=10)     9.926k ± 0%  -0.01% (p=0.022 n=10)
geomean                                               3.376k        3.376k       +0.00%                    3.375k       -0.01%
¹ all samples are equal

                                                  │ interface.txt │             generic.txt             │
                                                  │     B/op      │     B/op      vs base               │
StoreListCreate/RV=NotOlderThan-10                   213.7Ki ± 3%   203.9Ki ± 4%  -4.58% (p=0.000 n=10)
StoreListCreate/RV=ExactMatch-10                     116.0Ki ± 0%   116.1Ki ± 0%       ~ (p=0.218 n=10)
StoreList/RV=Empty-10                                6.023Mi ± 0%   6.023Mi ± 0%       ~ (p=0.912 n=10)
StoreList/RV=NotOlderThan-10                         6.007Mi ± 0%   6.007Mi ± 0%       ~ (p=0.810 n=10)
StoreList/RV=MatchExact-10                           60.34Mi ± 0%   60.35Mi ± 0%       ~ (p=0.579 n=10)
StoreList/Paginate/RV=Empty-10                       64.07Mi ± 0%   64.12Mi ± 0%  +0.08% (p=0.000 n=10)
StoreList/Paginate/RV=NotOlderThan-10                64.04Mi ± 0%   64.11Mi ± 0%  +0.11% (p=0.000 n=10)
StoreList/Paginate/RV=MatchExact-10                  63.56Mi ± 0%   63.62Mi ± 0%  +0.10% (p=0.000 n=10)
StoreList/NodeIndexed/RV=Empty-10                    7.886Mi ± 1%   7.931Mi ± 1%  +0.58% (p=0.009 n=10)
StoreList/NodeIndexed/RV=NotOlderThan-10             6.278Mi ± 0%   6.321Mi ± 0%  +0.68% (p=0.000 n=10)
StoreList/NodeIndexed/RV=MatchExact-10               5.779Gi ± 0%   5.782Gi ± 0%       ~ (p=0.052 n=10)
StoreList/NodeIndexed/Paginate/RV=Empty-10           250.5Mi ± 0%   255.6Mi ± 0%  +2.02% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=NotOlderThan-10    248.8Mi ± 0%   253.9Mi ± 0%  +2.04% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=MatchExact-10      6.039Gi ± 0%   6.050Gi ± 0%  +0.19% (p=0.000 n=10)
StoreList/Namespace/RV=Empty-10                      8.136Mi ± 0%   8.140Mi ± 0%  +0.05% (p=0.007 n=10)
StoreList/Namespace/RV=NotOlderThan-10               6.529Mi ± 0%   6.529Mi ± 0%       ~ (p=0.529 n=10)
StoreList/Namespace/RV=MatchExact-10                 63.30Mi ± 0%   63.36Mi ± 0%  +0.10% (p=0.000 n=10)
geomean                                              28.41Mi        28.43Mi       +0.08%

                                                  │ interface.txt │            generic.txt             │
                                                  │   allocs/op   │  allocs/op   vs base               │
StoreListCreate/RV=NotOlderThan-10                     558.0 ± 0%    560.0 ± 0%  +0.36% (p=0.000 n=10)
StoreListCreate/RV=ExactMatch-10                       868.5 ± 0%    875.0 ± 0%  +0.75% (p=0.000 n=10)
StoreList/RV=Empty-10                                  351.0 ± 2%    349.0 ± 4%       ~ (p=0.616 n=10)
StoreList/RV=NotOlderThan-10                           88.00 ± 0%    88.00 ± 1%       ~ (p=1.000 n=10)
StoreList/RV=MatchExact-10                            447.3k ± 0%   450.2k ± 0%  +0.64% (p=0.000 n=10)
StoreList/Paginate/RV=Empty-10                        478.6k ± 0%   481.6k ± 0%  +0.63% (p=0.000 n=10)
StoreList/Paginate/RV=NotOlderThan-10                 478.4k ± 0%   481.4k ± 0%  +0.63% (p=0.000 n=10)
StoreList/Paginate/RV=MatchExact-10                   480.5k ± 0%   483.9k ± 0%  +0.70% (p=0.000 n=10)
StoreList/NodeIndexed/RV=Empty-10                     32.20k ± 0%   32.40k ± 0%  +0.61% (p=0.000 n=10)
StoreList/NodeIndexed/RV=NotOlderThan-10              6.732k ± 1%   6.734k ± 1%       ~ (p=0.867 n=10)
StoreList/NodeIndexed/RV=MatchExact-10                45.64M ± 0%   45.92M ± 0%  +0.63% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=Empty-10            1.797M ± 0%   1.889M ± 0%  +5.16% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=NotOlderThan-10     1.769M ± 0%   1.861M ± 0%  +5.22% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=MatchExact-10       46.40M ± 0%   46.77M ± 0%  +0.79% (p=0.000 n=10)
StoreList/Namespace/RV=Empty-10                       31.80k ± 0%   32.01k ± 0%  +0.66% (p=0.000 n=10)
StoreList/Namespace/RV=NotOlderThan-10                6.333k ± 0%   6.333k ± 0%       ~ (p=0.444 n=10)
StoreList/Namespace/RV=MatchExact-10                  477.9k ± 0%   481.1k ± 0%  +0.68% (p=0.000 n=10)
geomean                                               76.42k        77.17k       +0.98%

                                                  │ interface.txt │             generic.txt             │            generics.txt             │
                                                  │   pages/op    │  pages/op   vs base                 │  pages/op   vs base                 │
StoreList/RV=Empty-10                                  1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹   1.000 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/RV=NotOlderThan-10                           1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹   1.000 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/RV=MatchExact-10                             1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹   1.000 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Paginate/RV=Empty-10                         100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Paginate/RV=NotOlderThan-10                  100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Paginate/RV=MatchExact-10                    100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/RV=Empty-10                      100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/RV=NotOlderThan-10               100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/RV=MatchExact-10                 100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/NodeIndexed/Paginate/RV=Empty-10             145.0 ± 0%   150.0 ± 0%  +3.45% (p=0.000 n=10)     146.0 ± 0%  +0.69% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=NotOlderThan-10      145.0 ± 0%   150.0 ± 0%  +3.45% (p=0.000 n=10)     146.0 ± 0%  +0.69% (p=0.000 n=10)
StoreList/NodeIndexed/Paginate/RV=MatchExact-10        145.0 ± 0%   148.0 ± 0%  +2.07% (p=0.000 n=10)     143.0 ± 0%  -1.38% (p=0.000 n=10)
StoreList/Namespace/RV=Empty-10                        100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Namespace/RV=NotOlderThan-10                 100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
StoreList/Namespace/RV=MatchExact-10                   100.0 ± 0%   100.0 ± 0%       ~ (p=1.000 n=10) ¹   100.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                42.88        43.13       +0.59%                    42.88       -0.00%
¹ all samples are equal

```